aci_tasks/requirements.txt
adal==1.2.2
attrs==19.3.0
azure-common==1.1.23
azure-mgmt-containerinstance==1.5.0
azure-mgmt-resource==6.0.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
entrypoints==0.3
flake8==3.7.9
idna==2.8
importlib-metadata==0.23
isodate==0.6.0
mccabe==0.6.1
more-itertools==7.2.0
msrest==0.6.10
msrestazure==0.6.2
oauthlib==3.1.0
packaging==19.2
pluggy==0.13.1
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
